Adding comments and double checking code
This USART example is still not working. I'm not sure about the baud
rate and perhaps that is the problem. In the debugger I've checked the
values and I can see that the character is written to the USART_DTR but
it is never removed which I expected it to be.