Remove TSC from ADC example
This commit removes the TSC code from this example as it adds more
complexity and was more of a misunderstanding on my part as to what the
TSC is supposed to be used for. I was focusing on the sensing part of
the name and through it would be able to use it as the source for the
analog to digital conversion. While it might be possible to use the
capacitor for that it still seems a little strange to do so. Instead
I've ordered a few sensors which I'll use for this example instead.