Extract uart transmit into separate object
This commit extract the uart init function and the write function so
that it can be used in other programs.

The motivation for doing this is to to be able to use these functions
for sending data from progams exploring SPI, I²C and CAN. The idea is
that when those programs receive/read data they can then pass that data
along to the setup UART and it can be displayed using minicom.