Add working CAN controller/peripheral example
The code needs to be cleaned up and that will happen in a followup
commit. This is currently working and I'm able to transmit using the
controller and the peripheral can receive and pass the message, a single
uppercase A and the identifier (7 in this example) using UART to be
inspected using minicom.